requirements.txt
fastapi
uvicorn
openai
httpx
pydub
numpy
google-cloud-speech
twilio
